+++ projects/ui-lib/src/lib/components/shared/ui-icon/index.ts
@@ -1,0 +1,2 @@
+export * from './ui-icon.component';
+export * from './ui-icon.interface';