+++ projects/ui-lib/src/lib/directives/index.ts
@@ -1,1 +1,2 @@
 export * from './link-type.directive';
+export * from './swiper.directive';