+++ projects/ui-lib/src/lib/modules/icons/icon.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { LucideAngularModule, ArrowUpRight, ArrowRight, ArrowLeft } from 'lucide-angular';
+
+const icons = {
+  ArrowUpRight,
+  ArrowRight,
+  ArrowLeft, 
+};
+
+@NgModule({
+  imports: [
+    LucideAngularModule.pick(icons)
+  ],
+  exports: [LucideAngularModule]
+})
+export class IconsModule {}