+++ projects/ui-lib/src/lib/services/mapper/component-mappers/social-footer.mapper.ts
@@ -1,0 +1,8 @@
+import { mapButtons, mapImage } from '../../../utils/utils';
+import { ComponentMapperFn } from './types';
+
+export const mapSocialFooter: ComponentMapperFn = (props) => ({ 
+    images: mapImage(props?.multimedia),
+    socials: mapButtons(props?.social_links),
+    hashtag: props?.texts?.hashtag,
+});